microservices/video_lecture_agent/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-genai
requests
pydantic
